restructuring the project

+++ src/controller/blogController.js
@@ -35,15 +35,15 @@ module.exports.searchBlog = async (req, res) => {
 module.exports.postBlog = async (req, res) => {
   try {
     const fileUpload = await cloudinary.uploader.upload(req.file.path);
-    // const { role } = req.user;
+    const { role } = req.user;
     const { blog_title, blog_description } = req.body;
     const blog_image = fileUpload.secure_url;
     if (!blog_title || !blog_description || !blog_image) {
       return res.json({ message: "all fields are required" });
     }
-    // if (role != "admin") {
-    //   return res.json({ message: "you are not allowed to post blog" });
-    // }
+    if (role != "admin") {
+      return res.json({ message: "you are not allowed to post blog" });
+    }
     const blog = {
       blog_title: blog_title,
       blog_image: blog_image,
@@ -59,7 +59,7 @@ module.exports.postBlog = async (req, res) => {
 
 module.exports.updateBlog = async (req, res) => {
   try {
-    // const { role } = req.user;
+    const { role } = req.user;
     const id = req.params.id;
     const { blog_title, blog_description } = req.body;
     const fileUpload = await cloudinary.uploader.upload(req.file.path);
@@ -67,9 +67,9 @@ module.exports.updateBlog = async (req, res) => {
     if (!blog_title || !blog_description || !id || !blog_image) {
       return res.json({ message: "all fields are required" });
     }
-    // if (role != "admin") {
-    //   return res.json({ message: "you are not allowed to update blog" });
-    // }
+    if (role != "admin") {
+      return res.json({ message: "you are not allowed to update blog" });
+    }
     const blog = await Blog.findById(id);
     blog.blog_title = blog_title || blog.blog_title;
     blog.blog_image = blog_image || blog.blog_image;
@@ -86,14 +86,14 @@ module.exports.updateBlog = async (req, res) => {
 
 module.exports.deleteBlog = async (req, res) => {
   try {
-    // const { role } = req.user;
+    const { role } = req.user;
     const id = req.params.id;
     if (!id) {
       return res.json({ message: "id is not provided" });
     }
-    // if (role != "admin") {
-    //   return res.json({ message: "you are not allowed to delete blog" });
-    // }
+    if (role != "admin") {
+      return res.json({ message: "you are not allowed to delete blog" });
+    }
     const blog = await Blog.findByIdAndDelete(id);
     return res.json({ message: "blog deleted sucessfully" }).status(200);
   } catch (err) {
